requirements.txt
feedparser>=6.0.10
requests>=2.31.0
beautifulsoup4>=4.12.0
anthropic>=0.3.0
python-dotenv>=1.0.0
jinja2>=3.1.0
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
